etl/requirements.txt
celery>=5.0.5
requests>=2.25.1
tika>=1.24
feedparser>=6.0.2
boto3>=1.17.0
psycopg2-binary>=2.8.6